+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * [#339 IMDB ID not available for movies](https://github.com/WebTools-NG/WebTools-NG/issues/339)
 * [#342 Rename "Export Arts"](https://github.com/WebTools-NG/WebTools-NG/issues/342)
 * [#345 Switch ET Export type selection](https://github.com/WebTools-NG/WebTools-NG/issues/345)
+* [#347 No Media Library is showing](https://github.com/WebTools-NG/WebTools-NG/issues/347)
 
 ## V0.3.4